Implementation of the backtracking algorithm
Two improvements:
1) unrolled_loops.c
The loops to find if a value exists in a row/column/box have been unrolled (to  avoid increasing counters and branching), based on AVR4027-Section 4.3

2) unrolled_MACROS.c
The functions to find if a value exists in a row/column/box have been converted to macros (with unrolled loops), to avoid hundreds/thousands of function calls. Based on AVR035, page 20.

/* By executing main, we can see how many times each function was called */